- Detalhando funcionários; - Correção nas exceções e regras de negócio; - Revisão campos obrigatórios;

+++ src/main/java/com/challenge/gestorprojetos/service/ProjetoServiceImpl.java
@@ -2,12 +2,12 @@ package com.challenge.gestorprojetos.service;
 
 import com.challenge.gestorprojetos.exception.ProjetoNaoEncontradoException;
 import com.challenge.gestorprojetos.model.Projeto;
+import com.challenge.gestorprojetos.model.Status;
 import com.challenge.gestorprojetos.repository.ProjetoRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.NoSuchElementException;
 
 @Service
 public class ProjetoServiceImpl implements ProjetoService {
@@ -32,11 +32,18 @@ public class ProjetoServiceImpl implements ProjetoService {
 
     @Override
     public void salvarProjeto(Projeto projeto) {
+        if (projeto.getGerenteResponsavel() == null) {
+            throw new IllegalArgumentException("É necessário informar um gerente para o projeto");
+        }
         projetoRepository.save(projeto);
     }
 
     @Override
     public void excluirProjeto(Projeto projeto) {
+        if (projeto.getStatus() == Status.INICIADO || projeto.getStatus() == Status.EM_ANDAMENTO ||
+                projeto.getStatus() == Status.ENCERRADO) {
+            throw new IllegalArgumentException("Não é possível excluir projeto com o status: " + projeto.getStatus().getDescricao());
+        }
         projetoRepository.delete(projeto);
     }
 }
